refactor: create new file for login querries and modularized, remove old testing code.
